feat: visually stop hand mesh in grasping pose
This is obtained by a set of patches to Meta SDK to access private class
data, and by creating a clonse of OVRHand just with rendering
components when a grasping action starts and disabling rendering
components on the original hand. Then restoring only the original hand
when grasping ends.

To reimplement this changes, replace Meta's OVRHand and OVRSkeleton with
the ones in this repo (either in the PackageCache) or directly into
UnityEditor.