unity/hand-controller: use fingerMixer weight as source for closure data for IK
The WeART SDK (in particular in the re-added WeArtHandController.AnimateFingers() ) takes care of different hand states to manipulate the animation (eg. disable when grasping, slow down when in vicinity of an object).
Furthermore, some grasping conditions (ConditionB(), ConditionC() in WeArtHandGraspingSystem) depend on the animation value.
Disable the Animator and the whole animation pipeline in WeArtHandController, but leave _fingerMixers and extracts weights from them.
Use these weights as source of closure and abduction data instead of the raw values for our animation using IK in FixedUpdate().